# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KakaoRelay Local Gateway",
    "version": "0.7.5",
    "description": "인증된 로컬 Windows 게이트웨이. 메시지 전송, SQLCipher 4 로컬 대화 읽기, CLI AI 분석·페르소나 설정."
  },
  "security": [
    {
      "localBearer": []
    }
  ],
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "지원 기능",
        "responses": {
          "200": {
            "description": "처리 결과. HTTP 200만으로 발송 완료를 판단하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "400": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rooms": {
      "get": {
        "operationId": "listRooms",
        "summary": "열린 대화방 조회",
        "responses": {
          "200": {
            "description": "처리 결과. HTTP 200만으로 발송 완료를 판단하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          },
          "400": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/send": {
      "post": {
        "operationId": "sendMessage",
        "summary": "앱을 통해 메시지 발송",
        "description": "같은 requestId와 같은 본문은 기존 결과 반환. 이름이 정확히 일치하는 열린 방 하나만 허용. 자동 재전송 없음.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "처리 결과. HTTP 200만으로 발송 완료를 판단하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests": {
      "get": {
        "operationId": "listRequests",
        "summary": "발송 기록 조회 (본문 제외)",
        "responses": {
          "200": {
            "description": "처리 결과. HTTP 200만으로 발송 완료를 판단하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Receipt"
                  }
                }
              }
            }
          },
          "400": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests/{requestId}": {
      "get": {
        "operationId": "getRequest",
        "summary": "요청별 결과 조회",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "처리 결과. HTTP 200만으로 발송 완료를 판단하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "기록 없음은 미발송의 증거가 아닙니다. 같은 요청 ID를 유지하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/local/rooms": {
      "get": {
        "operationId": "getLocalRooms",
        "summary": "카카오톡 로컬 방 새로고침",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalRoom"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 설정 또는 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "409": {
            "description": "키 미확인, DB 변경 중 또는 AI 작업 중"
          },
          "503": {
            "description": "AI 프로바이더 실행 실패"
          }
        }
      }
    },
    "/v1/local/messages": {
      "get": {
        "operationId": "getLocalMessages",
        "summary": "최근 로컬 메시지",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatContext"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 설정 또는 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "409": {
            "description": "키 미확인, DB 변경 중 또는 AI 작업 중"
          },
          "503": {
            "description": "AI 프로바이더 실행 실패"
          }
        },
        "parameters": [
          {
            "name": "profile",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ]
      }
    },
    "/v1/ai/settings": {
      "get": {
        "operationId": "getAiSettings",
        "summary": "AI 설정 조회",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettings"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 설정 또는 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "409": {
            "description": "키 미확인, DB 변경 중 또는 AI 작업 중"
          },
          "503": {
            "description": "AI 프로바이더 실행 실패"
          }
        }
      },
      "put": {
        "operationId": "saveAiSettings",
        "summary": "AI 설정 전체 저장",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "저장 완료"
          },
          "400": {
            "description": "설정 오류"
          }
        }
      }
    },
    "/v1/ai/providers": {
      "get": {
        "operationId": "getAiProviders",
        "summary": "CLI 설치 상태 조회",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CliStatus"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 설정 또는 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "409": {
            "description": "키 미확인, DB 변경 중 또는 AI 작업 중"
          },
          "503": {
            "description": "AI 프로바이더 실행 실패"
          }
        }
      }
    },
    "/v1/ai/generate": {
      "post": {
        "operationId": "generateAiReply",
        "summary": "분석·질문·답변 생성 (발송 없음)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiResult"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 설정 또는 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "409": {
            "description": "키 미확인, DB 변경 중 또는 AI 작업 중"
          },
          "503": {
            "description": "AI 프로바이더 실행 실패"
          }
        }
      }
    },
    "/v1/send-image": {
      "post": {
        "operationId": "sendImage",
        "summary": "이미지 1장 첨부·전송 (Windows 핸들, Computer Use 없음)",
        "description": "같은 requestId와 같은 본문은 기존 결과 반환. 이름이 정확히 일치하는 열린 방 하나만 허용. 자동 재전송 없음.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "처리 결과. HTTP 200만으로 발송 완료를 판단하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "오류 code/detail을 확인하세요.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/{personaId}/images": {
      "get": {
        "operationId": "listPersonaImages",
        "summary": "페르소나의 Astra/Grok 이미지와 적용 상태 조회 (읽기 전용)",
        "parameters": [
          {
            "name": "personaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "personaId, appliedProvider, versions: {astra, grok}. 각 버전의 referenceId, active, images."
          },
          "404": {
            "description": "페르소나 없음"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "localBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "앱이 매 실행 생성하는 connection.json의 token. 외부에 출력하지 마세요."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "SendCommand": {
        "type": "object",
        "required": [
          "requestId",
          "recipient",
          "message"
        ],
        "properties": {
          "requestId": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "description": "동일 작업의 재시도에는 동일 ID"
          },
          "recipient": {
            "type": "string",
            "description": "rooms의 정확한 title"
          },
          "message": {
            "type": "string",
            "description": "본문. 줄바꿈 지원. 임의 글자 수 제한 없음. 공백뿐인 본문 및 NUL 금지."
          },
          "sendDelayMs": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30000,
            "default": 1000,
            "description": "텍스트 입력 응답 후 Enter 전 대기(ms). 입력 응답 시간 초과에도 대기 후 진행하며 본문 재확인은 하지 않음. 이미지에는 미적용. 같은 요청 ID의 대기 시간을 바꿔도 재전송하지 않음."
          },
          "imagePath": {
            "type": "string",
            "description": "PC 로컬 이미지 절대 경로. PNG/JPEG/GIF/BMP/WebP."
          },
          "attachment": {
            "$ref": "#/components/schemas/PersonaAttachment"
          }
        }
      },
      "Room": {
        "type": "object",
        "required": [
          "title",
          "processId",
          "handle",
          "ambiguous",
          "selectable",
          "hint"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "processId": {
            "type": "integer"
          },
          "handle": {
            "type": "string"
          },
          "ambiguous": {
            "type": "boolean"
          },
          "selectable": {
            "type": "boolean"
          },
          "hint": {
            "type": "string"
          }
        }
      },
...
      "Capabilities": {
        "type": "object",
        "properties": {
          "send": {
            "const": true
          },
          "idempotency": {
            "const": "requestId"
          },
          "requiresOpenRoom": {
            "const": true
          },
          "maxMessageCharacters": {
            "type": "null"
          },
          "recentMessages": {
            "type": "object",
            "properties": {
              "supported": {
                "type": "boolean"
              },
              "reason": {
                "type": "string"
              }
            }
          },
          "deliveryVerification": {
            "const": "manual"
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aiChat": {
            "type": "boolean"
          }
        }
      },
      "LocalRoom": {
        "type": "object",
        "required": [
          "profile",
          "id",
          "title",
          "readable"
        ],
        "properties": {
          "profile": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "readable": {
            "type": "boolean"
          }
        }
      },
...
      "ChatContext": {
        "type": "object",
        "required": [
          "room",
          "messages",
          "readAt"
        ],
        "properties": {
          "room": {
            "$ref": "#/components/schemas/LocalRoom"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalMessage"
            }
          },
          "readAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CliStatus": {
        "type": "object",
        "required": [
          "provider",
          "installed",
          "path",
          "detail"
        ],
        "properties": {
          "provider": {
            "type": "string"
          },
          "installed": {
            "type": "boolean"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "PersonaAttachment": {
        "type": "object",
        "required": [
          "personaId",
          "provider",
          "imageId"
        ],
        "properties": {
          "personaId": {
            "type": "string"
          },
          "provider": {
            "enum": [
              "astra",
              "grok"
            ]
          },
          "imageId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{32}$"
          }
        }
      },
...
    }
  }
}